Move post functionality in post_resolver
