+++ lib/sanbase_web/graphql/middlewares/post_permissions.ex
@@ -3,7 +3,7 @@ defmodule SanbaseWeb.Graphql.Middlewares.PostPermissions do
 
   alias Absinthe.Resolution
 
-  @allowed_fields_for_anon_users ["title", "short_desc"]
+  @allowed_fields_for_anon_users ["title", "shortDesc"]
 
   def call(%Resolution{context: %{auth: %{auth_method: :user_token}}} = resolution, _) do
     resolution
